# swagger.yaml
openapi: 3.0.0
info:
  title: Workout Management API
  description: A complete RESTful API backend for a gym workout tracking application built with Node.js, Express, and PostgreSQL.
  version: 1.0.0
  contact:
    email: support@workoutapi.com
servers:
  - url: http://localhost:3000
    description: Local development server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Workout:
      type: object
      properties:
        id:
          type: integer
          example: 1
        id_user:
          type: integer
          example: 1
        name:
          type: string
          example: "Push Day"
        description:
          type: string
          nullable: true
          example: "Chest, shoulders and triceps"
        image_url:
          type: string
          nullable: true
          example: "/images/workouts/push-day.jpg"
        exercises:
          type: array
          items:
            $ref: '#/components/schemas/WorkoutExercise'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    WorkoutExercise:
      type: object
      properties:
        id:
          type: integer
        id_exercise:
          type: integer
        sets:
          type: integer
          example: 3
        number:
          type: integer
          example: 1
        name:
          type: string
          example: "Bench Press"
        muscle_group:
          type: string
          example: "Chest"
...
    RoutineWorkout:
      type: object
      properties:
        id:
          type: integer
        id_workout:
          type: integer
        number:
          type: integer
          example: 1
        name:
          type: string
          example: "Push Day"
        description:
          type: string
          nullable: true
...
paths:
...
  /api/workouts:
    get:
      tags:
        - Workouts
      summary: Get all user workouts
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Workouts retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  workouts:
                    type: array
                    items:
                      $ref: '#/components/schemas/Workout'
    post:
      tags:
        - Workouts
      summary: Create new workout
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                  example: "Push Day"
                description:
                  type: string
                  example: "Chest, shoulders and triceps"
                image:
                  type: string
                  format: binary
      responses:
        '201':
          description: Workout created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Workout created successfully"
                  workout:
                    $ref: '#/components/schemas/Workout'
  /api/workouts/{id}:
    get:
      tags:
        - Workouts
      summary: Get workout with exercises
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            example: 1
      responses:
        '200':
          description: Workout retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  workout:
                    $ref: '#/components/schemas/Workout'
    put:
      tags:
        - Workouts
      summary: Update workout
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            example: 1
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                image:
                  type: string
                  format: binary
      responses:
        '200':
          description: Workout updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Workout updated successfully"
                  workout:
                    $ref: '#/components/schemas/Workout'
    delete:
      tags:
        - Workouts
      summary: Delete workout
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            example: 1
      responses:
        '200':
          description: Workout deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Workout deleted successfully"
  /api/workouts/{id}/exercises:
    post:
      tags:
        - Workouts
      summary: Add exercise to workout
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - exercise_id
                - sets
                - number
              properties:
                exercise_id:
                  type: integer
                  example: 1
                sets:
                  type: integer
                  example: 3
                number:
                  type: integer
                  example: 1
      responses:
        '201':
          description: Exercise added to workout successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Exercise added to workout successfully"
                  workout_exercise:
                    $ref: '#/components/schemas/WorkoutExercise'
  /api/workouts/{workoutId}/exercises/{exerciseId}:
    delete:
      tags:
        - Workouts
      summary: Remove exercise from workout
      security:
        - BearerAuth: []
      parameters:
        - name: workoutId
          in: path
          required: true
          schema:
            type: integer
            example: 1
        - name: exerciseId
          in: path
          required: true
          schema:
            type: integer
            example: 1
      responses:
        '200':
          description: Exercise removed from workout successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Exercise removed from workout successfully"
  /api/workouts/{workoutId}/exercises/{exerciseId}/order:
    put:
      tags:
        - Workouts
      summary: Update exercise order in workout
      security:
        - BearerAuth: []
      parameters:
        - name: workoutId
          in: path
          required: true
          schema:
            type: integer
            example: 1
        - name: exerciseId
          in: path
          required: true
          schema:
            type: integer
            example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - number
              properties:
                number:
                  type: integer
                  example: 2
      responses:
        '200':
          description: Exercise order updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Exercise order updated successfully"
                  workout_exercise:
                    $ref: '#/components/schemas/WorkoutExercise'
...
  /api/routines/{id}/workouts:
    post:
      tags:
        - Routines
      summary: Add workout to routine
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - workout_id
                - number
              properties:
                workout_id:
                  type: integer
                  example: 1
                number:
                  type: integer
                  example: 1
      responses:
        '201':
          description: Workout added to routine successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Workout added to routine successfully"
                  routine_workout:
                    $ref: '#/components/schemas/RoutineWorkout'
  /api/routines/{routineId}/workouts/{workoutId}:
    delete:
      tags:
        - Routines
      summary: Remove workout from routine
      security:
        - BearerAuth: []
      parameters:
        - name: routineId
          in: path
          required: true
          schema:
            type: integer
            example: 1
        - name: workoutId
          in: path
          required: true
          schema:
            type: integer
            example: 1
      responses:
        '200':
          description: Workout removed from routine successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Workout removed from routine successfully"
...